Create fields module to hold custom fields
