+++ CTFd/forms/email.py
@@ -1,7 +1,8 @@
 from wtforms import TextAreaField
 from wtforms.validators import InputRequired
 
-from CTFd.forms import BaseForm, SubmitField
+from CTFd.forms import BaseForm
+from CTFd.forms.fields import SubmitField
 
 
 class SendEmailForm(BaseForm):
