+++ CTFd/forms/submissions.py
@@ -1,7 +1,8 @@
 from wtforms import SelectField, StringField
 from wtforms.validators import InputRequired
 
-from CTFd.forms import BaseForm, SubmitField
+from CTFd.forms import BaseForm
+from CTFd.forms.fields import SubmitField
 
 
 class SubmissionSearchForm(BaseForm):
