+++ CTFd/forms/users.py
@@ -2,7 +2,8 @@ from wtforms import BooleanField, PasswordField, SelectField, StringField
 from wtforms.fields.html5 import EmailField
 from wtforms.validators import InputRequired
 
-from CTFd.forms import BaseForm, SubmitField
+from CTFd.forms import BaseForm
+from CTFd.forms.fields import SubmitField
 from CTFd.utils.countries import SELECT_COUNTRIES_LIST
